+++ src/components/PreviousSummits/SummitCardLight/index.jsx
@@ -6,7 +6,7 @@ import {wrapper, image, title, subtitle, button} from './styles.module.scss';
 const SummitCardLight = ({summit}) => {
   return (
     <div className={wrapper}>
-      <StaticImage className={image} src={summit.image} alt={summit.name} />
+      {/*<StaticImage className={image} src={summit.image} alt={summit.name} />*/}
       <p className={title}>{summit.name}</p>
       <p className={subtitle}>{summit.date}</p>
       <a href={summit.link} className={button}>See Highlights --> </a>
