+++ src/pages/events/berlin-2022/index.md
@@ -22,7 +22,7 @@ header:
   buttons:
     - text: Watch Summit Videos
       link: https://www.openstack.org/videos/summits/berlin-2022
-  image: /img/summit/summit-hero-image.png
+  image: img/summit/summit-hero-image.png
 form:
   display: true
   title: Subscribe to our newsletter for the latest from the OpenInfra Foundation.
