+++ src/templates/berlin-2022-summit-page.js
@@ -1,4 +1,4 @@
-import React from 'react'
+  import React from 'react'
 import PropTypes from 'prop-types'
 import { graphql } from 'gatsby';
 import Layout from '../components/Layout'
